requirements.txt
telethon
openai
pandas
openpyxl
streamlit
python-dotenv